delivery price accept float

+++ api/src/app/models/Menu.js
@@ -4,7 +4,7 @@ class Menu extends Model {
   static init(sequelize) {
     super.init({
       description: DataTypes.STRING,
-      delivery_price: DataTypes.INTEGER,
+      delivery_price: DataTypes.FLOAT,
       logo_path: DataTypes.STRING,
       banner_path: DataTypes.STRING,
     }, {
